refactor: update KYC verification structure and remove deprecated files
- Changed the environment configuration to use SERVER_URL instead of HOST_DOMAIN.
- Updated the KYC verification request structure to use types from the new shared types package.
- Removed the old KYC types and configuration files, streamlining the KYC service implementation.
- Adjusted tests to reflect the new structure and ensure compatibility with the updated KYC provider interface.
