+++ .env.example
@@ -4,12 +4,12 @@ DEBUG=True
 ALLOWED_HOSTS=0.0.0.0
 SERVER_HOST=0.0.0.0
 SERVER_PORT=8000
+SERVER_URL=http://localhost:8000
 JWT_ACCESS_LIFESPAN=15
 JWT_REFRESH_LIFESPAN=10080
 HMAC_TIMESTAMP_AGE=5
 ENVIRONMENT=local # local, staging, production
 SENTRY_DSN=
-HOST_DOMAIN=http://localhost:8000
 
 # Database Config
 DB_NAME=paycrest
