--- services/kyc/smile/config/id_types.go
@@ -1,72 +1,0 @@
-package config
-
-import (
-	"encoding/json"
-	"fmt"
-	"os"
-
-	"github.com/xeipuuv/gojsonschema"
-)
-
-type IDType struct {
-	Type               string `json:"type"`
-	VerificationMethod string `json:"verification_method"`
-}
-
-type Country struct {
-	Name    string   `json:"name"`
-	Code    string   `json:"code"`
-	IDTypes []IDType `json:"id_types"`
-}
-
-type Continent struct {
-	Name      string    `json:"name"`
-	Countries []Country `json:"countries"`
-}
-
-type SmileIDConfig struct {
-	Continents []Continent `json:"continents"`
-}
-
-// SmileIDWebhookPayload represents the payload structure from Smile Identity
-type SmileIDWebhookPayload struct {
-	ResultCode    string `json:"ResultCode"`
-	PartnerParams struct {
-		UserID string `json:"user_id"`
-	} `json:"PartnerParams"`
-	Signature string `json:"signature"`
-	Timestamp string `json:"timestamp"`
-	// Add other fields as needed
-}
-
-func ValidateSmileIDConfig(filePath string) error {
-	// Read the config file
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		return fmt.Errorf("failed to read config file: %w", err)
-	}
-
-	// Validate against JSON schema
-	schemaPath := "./id_types_schema.json"
-	schemaLoader := gojsonschema.NewReferenceLoader("file://" + schemaPath)
-	documentLoader := gojsonschema.NewBytesLoader(data)
-	result, err := gojsonschema.Validate(schemaLoader, documentLoader)
-	if err != nil {
-		return fmt.Errorf("failed to validate schema: %w", err)
-	}
-	if !result.Valid() {
-		var errors []string
-		for _, e := range result.Errors() {
-			errors = append(errors, e.String())
-		}
-		return fmt.Errorf("schema validation failed: %v", errors)
-	}
-
-	// Ensure JSON is parseable into SmileIDConfig
-	var config SmileIDConfig
-	if err := json.Unmarshal(data, &config); err != nil {
-		return fmt.Errorf("failed to parse JSON: %w", err)
-	}
-
-	return nil
-}
