+++ services/kyc/smile/id_types_schema.json
