+++ utils/test/test.env
@@ -1,5 +1,5 @@
 SECRET=h9wt*pasj6796jw(w8=xaje8tpi6+k2)
-HOST_DOMAIN=https://example.com
+SERVER_URL=https://example.com
 
 # Test Database Config
